efficient broadcasting
To an op-assignment like '+=' between the left hand side operand A and the right hand side operand B, which is broadcastable to A,
there are two possible ways to do broadcast.
The top-down method seperates A as fine as possible, and passes the elements by reference in a recursive way;
The bottom-up approach constructs higher-dimensional arrays from B to match with A.

array.hpp demonstrates how the top-down method, which is adopted in waffle, outperforms its bottom-up counterpart,
as it avoids unnecessary imtermediate variables.

integer.hpp contains some simple type aliasing.
intrinsics.hpp is copied from previous implementations, yet not tested.