better assembly generation | bugfix of intrinsics
waffle::array's underlying assignment loop is changed to a simpler way. Now if the loop body is only executed once, it will be optimized away.
waffle's intrinsics assignment had a foolish bug (the result is not stored).
Trying the paradigm of SOA. It seems that generics should be in the way of
   <template <typename> typename ... Containers>
   struct foo {};