the first step to cross-platform
CMakeLists.txt is added and the compilations under GNU and Clang are tested.

All the headers are moved into include/waffle. Source files are under root directory for now.

The name of template argument in some member functions of waffle::array<T> is changed to P, in order to avoid name conflict.

Division of integer packet is disabled for now because its implementation is not trivial.
Subscript operator is rewritten to support cross-platform.